Split airspace implementation to smaller modules
