+++ flights/src/core/airspace/mod.rs
@@ -1,0 +1,4 @@
+mod detail;
+mod task;
+
+pub use task::{AirspaceStore, AirspaceViewer};
